refactor(auth): move header component outside of auth directory
